requirements.txt
pinecone-client
langchain
transformers
pillow
pandas
requests
torch
groq
flask
requests